+++ models/buffer_disk.go
@@ -6,17 +6,18 @@ import (
 	"os"
 	"sync"
 
+	"github.com/tidwall/wal"
+
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/metric"
-	"github.com/tidwall/wal"
 )
 
 type DiskBuffer struct {
 	BufferStats
 	sync.Mutex
 
-	walFile     *wal.Log
-	walFilePath string
+	file *wal.Log
+	path string
 
 	batchFirst uint64 // Index of the first metric in the batch
 	batchSize  uint64 // Number of metrics currently in the batch
@@ -34,8 +35,8 @@ func NewDiskBuffer(name string, path string, stats BufferStats) (*DiskBuffer, er
 	}
 	return &DiskBuffer{
 		BufferStats: stats,
-		walFile:     walFile,
-		walFilePath: filePath,
+		file:        walFile,
+		path:        filePath,
 	}, nil
 }
 
@@ -55,7 +56,7 @@ func (b *DiskBuffer) length() int {
 
 // readIndex is the first index to start reading metrics from, or the head of the buffer
 func (b *DiskBuffer) readIndex() uint64 {
-	index, err := b.walFile.FirstIndex()
+	index, err := b.file.FirstIndex()
 	if err != nil {
 		panic(err) // can only occur with a corrupt wal file
 	}
@@ -64,7 +65,7 @@ func (b *DiskBuffer) readIndex() uint64 {
 
 // writeIndex is the first index to start writing metrics to, or the tail of the buffer
 func (b *DiskBuffer) writeIndex() uint64 {
-	index, err := b.walFile.LastIndex()
+	index, err := b.file.LastIndex()
 	if err != nil {
 		panic(err) // can only occur with a corrupt wal file
 	}
@@ -91,7 +92,7 @@ func (b *DiskBuffer) addSingle(m telegraf.Metric) bool {
 	if err != nil {
 		panic(err)
 	}
-	err = b.walFile.Write(b.writeIndex(), data)
+	err = b.file.Write(b.writeIndex(), data)
 	if err == nil {
 		b.metricAdded()
 		return true
@@ -112,7 +113,7 @@ func (b *DiskBuffer) addBatch(metrics []telegraf.Metric) int {
 		b.metricAdded()
 		written++
 	}
-	err := b.walFile.WriteBatch(batch)
+	err := b.file.WriteBatch(batch)
 	if err != nil {
 		return 0 // todo error handle, test if a partial write occur
 	}
@@ -125,7 +126,7 @@ func (b *DiskBuffer) Batch(batchSize int) []telegraf.Metric {
 
 	if b.length() == 0 {
 		// no metrics in the wal file, so return an empty array
-		return make([]telegraf.Metric, 0)
+		return []telegraf.Metric{}
 	}
 	b.batchFirst = b.readIndex()
 	var metrics []telegraf.Metric
@@ -134,7 +135,7 @@ func (b *DiskBuffer) Batch(batchSize int) []telegraf.Metric {
 	readIndex := b.batchFirst
 	endIndex := b.writeIndex()
 	for batchSize > 0 && readIndex < endIndex {
-		data, err := b.walFile.Read(readIndex)
+		data, err := b.file.Read(readIndex)
 		if err != nil {
 			panic(err)
 		}
@@ -175,7 +176,7 @@ func (b *DiskBuffer) Accept(batch []telegraf.Metric) {
 	if b.length() == len(batch) {
 		b.resetWalFile()
 	} else {
-		err := b.walFile.TruncateFront(b.batchFirst + uint64(len(batch)))
+		err := b.file.TruncateFront(b.batchFirst + uint64(len(batch)))
 		if err != nil {
 			panic(err)
 		}
@@ -211,11 +212,11 @@ func (b *DiskBuffer) resetBatch() {
 // todo to actually clear the walfile completely if needed, since Truncate() calls require
 // todo at least one entry remains in them otherwise they return an error.
 func (b *DiskBuffer) resetWalFile() {
-	b.walFile.Close()
-	os.Remove(b.walFilePath)
-	walFile, err := wal.Open(b.walFilePath, nil)
+	b.file.Close()
+	os.Remove(b.path)
+	walFile, err := wal.Open(b.path, nil)
 	if err != nil {
 		panic(err)
 	}
-	b.walFile = walFile
+	b.file = walFile
 }
